# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Apex UEEP API
  description: Universal Enterprise AI Operating System API. Monetize data, orchestrate AI agents, and manage billing.
  version: 1.0.0
  contact:
    email: gwc2780@gmail.com
servers:
  - url: https://apex-api.railway.app
    description: Production Server
  - url: http://localhost:8000
    description: Local Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    EdgeRevenueEvent:
      type: object
      required:
        - node_id
        - charge_id
        - amount_usd
        - plan
        - source
      properties:
        node_id:
          type: string
          example: "pixel-10-edge-001"
        charge_id:
          type: string
          example: "charge-edge-001"
        amount_usd:
          type: number
          format: float
          example: 149.0
        plan:
          type: string
          example: "pro"
        source:
          type: string
          example: "edge"
        timestamp:
          type: string
          format: date-time
security:
  - BearerAuth: []
paths:
...
  /webhook/edge/revenue:
    post:
      summary: Edge Revenue Event Webhook
      description: Receives revenue collection events from edge nodes and records them in the unified revenue ledger.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EdgeRevenueEvent'
      responses:
        '200':
          description: Revenue event recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "recorded"
        '400':
          description: Missing required fields or invalid amount_usd
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string